Add files via upload
Can natural language processing find the hidden-in-plain-sight data citations? Can machine learning find the link between the words used in research articles and the data referenced in the article? In this project use natural language processing (NLP) to automate the discovery of how scientific data are referenced in publications and labels